Day-13/src/requirements.txt
azure-functions
textgenrnn
tensorflow==1.13.1